requirements.txt
google-generativeai
youtube-transcript-api
pytubefix
python-dotenv
yt-dlp
openai-whisper
requests 